requirements/base.txt
django==4.2.5
djangorestframework==3.14.0
django-environ==0.11.2
psycopg2==2.9.7
ipython==8.16.1
djangorestframework-simplejwt==5.3.0